Added Model for Add Assignment
I added a model for the add assignment wire frame, previously had more but messed everything up and had to revert to previous version. 

AddEvent1 is no longer being used but is still there for the sake of me not wanting to delete anything because it caused problems previously. 

(Refer to comment on AddEvent1 for further information)