Fixed the Testing branch and added "AddPersonalEvent" model
It was not running, there was a view with no model, (forgot to delete it when getting rid of files related to old Event model). Code now runs.

AddPersonalEvent is here first because Some datetime stuff and a not required feature for one of the fields needs to be tested here so that main doesn't get messed up.