Added comment to Add Assignment
So basically add assignment needs to have a not required time filed added for there be the option to add specific times to an assignment. I will be adding that soon but just in case it is on a new branch to not mess up main, because the add assignment file isn't on testing. This branch should only be used for this specific task before it is merged back into main. Do not merge test into main before this branch. This task should be completed first.