AddPersonalEvents is done with working Non Required fields
Ok, so basically, the model for AddPersonalEvents is done. I've decided that personal events should be at least an hour so I used an int variable instead of the kind with a decimal. This is ready to be merged into main, but wanted to discuss with group first of whether or not some things are fine. 

There are some problems with add migration but should not be a future problem so nothing to worry about I guess.