Replaced AddAssignment with AddAssignment2
So basically AddAssignment needed some new fields and some changes and messing with the database makes things not work sometimes, so for the sake of simplicity I just made a new model that does the same thing but it's updated. On this commit it is not in the database yet, and I'm using this commit as a checkpoint so if it doesn't work, I can easily go back to make changes.