Added Members View Model
So the main change here has been the addition of a folder called members view models and a view model in said folders for the members class. 

THIS IS STILL A WORK IN PROGRESS

So basically I made a view model for the purpose of being used in the MembersController to add things to a list.

Why am I doing this?
This list will organize assignments of users. Things need to be added to that list in a controller. I am using last years FBLA backend to help me as I don't really know what I am doing. A view model was used in their controller to add things to their List, so I shall be doing something similar to theirs. 

What is the code (FBLA back end from last year) I am referring to?

The code I am referring to can be found in the images on this google doc:

https://docs.google.com/document/d/122G2z3KoAPsY9fS41RUvaPkLfC_RdjnwqSjV_It-U4g/edit?usp=sharing 

Also there are a few lines of commented out code and psudo code. Do not delete them. They are located on the new viewmodel, Members.cs, and MembersController.cs