Added Class Model
So this summary is for the creation of the new branch but it's in this commit because it didn't let me put it in the last commit. So instead of saving the assignments to the member, they will instead be saved to a class, as in a class with a teacher. This makes more sense so that everyone can see assignments such as homework, while a separate list is made for personal assignments, which get saved to the member. I added a class model, not yet scaffolded, for the classes (subject to change), and probably need to make a vm for the Class_ model (has underscore to not be mistaken with unnamed class).