Assignments, and date and times collected
So basically now there's a list called Ua that stores an unorganized list of a users assignments, as well as the datetime, timechoice, and other important date variables from the assignment model. The list is type UserAssignments, which is composed of the name of the assignment and other related dates and times.