Assignments linked to users via user Ids and some fixes to db context
So assignment were linked to the users via their user ids, this should also be done to the notes, future changes should include adding authentication to the pages (as in you can't edit things while not logged in). The user view model was in the dbcontext for some reason so it was removed because it should not be there.